finalized styles for buttons and controls
